dev: move /dev/* code to separate file
Devices can be enabled/disabled through lwipopts.h file:
	LWIP_ROUTE_DEV (1 by default)
	LWIP_IFSTATUS_DEV (1 by default)
	LWIP_EXT_PF
	LWIP_LINKMONITOR_DEV (0 by default)
/dev/route and /dev/ifstatus internal buffer sizes can be set by:
	LWIP_ROUTE_DEV_BUFFER_SIZE (1024 by default)
	LWIP_IFSTATUS_DEV_BUFFER_SIZE (512 by default)
Linkmonitor device can be set by:
	LWIP_LINKMONITOR_DEV_NAME (en1 by default)

JIRA: DTR-91
